Added form-based authentication to the redirector
Spring-security-core 2.0 RC4 was introduced to support authentication. The login page was adjusted in terms of styling. The CRUD operations for shorteners are locked down so that only authenticated users can see it. Additionally default users were created: admin:admin and user:user (with corresponding roles).

The functional test were adjusted so that they login as user:user before actually doing anything. Other tests had to be modified as well, so that the userCreated attribute of the shortener is no longer a string, but a reference to a real user instance insted.